submodule: explicitly specify on-demand upon push
When running "git push --recurse-submodules=on-demand" (or "=only") on a
superproject with nested submodules and assuming default configurations,
only the top-level submodules are pushed. This is because recursion
into the top-level submodules is performed with "git push" (without any
additional relevant options), and the default configuration is to not
recurse to nested submodules.

Therefore, instead of recursing with "git push" without any additional
relevant options, recurse with "--recurse-submodules=on-demand".

This now means that any push.recurseSubmodules configuration in any
submodule is no longer respected: only the configuration (or CLI
argument to override it) of the superproject is used. Update the
documentation accordingly.

(As a side effect of making the documentation of both the CLI argument
and the config variable consistent, the config variable is now
documented to support "only". This has been the case for a while, and I
have also included a test to show that.)

Signed-off-by: Jonathan Tan <jonathantanmy@google.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>